Add Bitrix content manager course to education section
Inserted code to include the Bitrix content manager course in the list of educational experiences on the website. The new list item is added under the "Education" section with a corresponding popup for more details. This enhances the user's understanding of the individual's education background by including information about the Bitrix content manager course.